langchain_prompts/messages.py
from pyexpat import model
from langchain_core.messages import SystemMessage, HumanMessage, AIMessage
from langchain_openai import ChatOpenAI
from dotenv import load_dotenv

load_dotenv()

model = ChatOpenAI(model='gpt-5-nano')

messages=[
    SystemMessage(content='You are a helpful assistant'),
    HumanMessage(content='tell me about Langchain')
]

result = model.invoke(messages)

messages.append(AIMessage(content=result.content))

print(messages)